add insident and rate

+++ rutaUQ-backend/src/main/resources/db/migration/V15__ratings.sql
@@ -2,7 +2,7 @@ CREATE TABLE ratings (
     id             UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     reservation_id UUID NOT NULL UNIQUE REFERENCES reservations(id),
     rater_id       UUID NOT NULL REFERENCES users(id),
-    stars          SMALLINT NOT NULL CHECK (stars BETWEEN 1 AND 5),
+    stars          INTEGER NOT NULL CHECK (stars BETWEEN 1 AND 5),
     comment        VARCHAR(500),
     created_at     TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now()
 );
